got it nice now

+++ bag.odin
@@ -2,14 +2,14 @@ package main
 import "core:math/rand"
 
 Bag :: struct {
-	size:  int,
-	array: [SUDOKU_SIZE]int,
+	size:  u8,
+	array: [SUDOKU_SIZE]u8,
 }
 
 // creating a tetris bag
 bag_new :: proc() -> (b: Bag) #no_bounds_check {
 	#unroll for x in 0 ..< SUDOKU_SIZE {
-		b.array[x] = x + 1
+		b.array[x] = u8(x + 1)
 	}
 	// fisher yates shuffle
 	// it is backwards on purpose, and also doesn't act on the index 0
@@ -22,24 +22,18 @@ bag_new :: proc() -> (b: Bag) #no_bounds_check {
 }
 
 // very crude append
-bag_append :: proc(bag: ^Bag, value: ^int) #no_bounds_check {
-	// log.info("append")
-	// log.info(bag)
+bag_append :: proc(bag: ^Bag, value: ^u8) #no_bounds_check {
 	bag.array[bag.size] = value^
 	bag.size += 1
-	// log.info(bag)
 }
 
 //very unsafe pop
-bag_pop_front :: proc(bag: ^Bag) -> (result: int) #no_bounds_check {
-	// log.info("pop front")
-	// log.info(bag)
+bag_pop_front :: proc(bag: ^Bag) -> (result: u8) #no_bounds_check {
 	result = bag.array[0]
 	if bag.size > 1 {
 		copy(bag.array[0:], bag.array[1:])
 	}
 	bag.size -= 1
 	bag.array[bag.size] = 0
-	// log.info(bag)
 	return
 }
